HTML-CSS
Create a movie ticket booking website using HTML and CSS. The website should have a homepage that allows users to choose between the
booking page and the now-showing page. The booking page should allow
users to select a movie to watch with several available and booked seats, and the now-showing page should display the different movies available for the show along with the time.